Rewrite all dataLoaders to the new composable style
The old approach was not really composable: now that we want to load
insights from both InsightView and DashboardView, the current way
is cumbersome to use.

The new way has several advantages:
* it is more granular, so that it can be composed better
* the loader caching is abstracted and tested only once

Also adds missing loaders for DashboardView.

JIRA: RAIL-2858
