+++ libs/sdk-ui-ext/.dependency-cruiser.js
@@ -7,6 +7,7 @@ const options = {
         // TODO: uncomment this line when dashboardEmbedding is no longer internal (for now this is to avoid circular dependencies)
         // depCruiser.isolatedSubmodule("internal", "src/internal"),
         depCruiser.moduleWithDependencies("insightView", "src/insightView", [
+            "src/dataLoaders",
             "src/internal",
             // TODO: remove this line when dashboardEmbedding is no longer internal (for now this is to avoid circular dependencies)
             "src/internal/*",
