+++ libs/sdk-ui-ext/src/dataLoaders/constants.ts
@@ -1,0 +1,11 @@
+// (C) 2021 GoodData Corporation
+
+/**
+ * Loaders for how many different workspaces will be kept.
+ */
+export const LOADER_CACHE_SIZE = 5;
+
+/**
+ * How many insights will be kept for each workspace.
+ */
+export const INSIGHT_CACHE_SIZE = 20;
