+++ libs/sdk-ui-ext/src/dataLoaders/types.ts
@@ -1,0 +1,8 @@
+// (C) 2021 GoodData Corporation
+/**
+ * @internal
+ */
+export interface IDataLoaderFactory<TLoader> {
+    forWorkspace(workspace: string): TLoader;
+    reset: () => void;
+}
