+++ libs/sdk-ui-ext/src/index.ts
@@ -1,3 +1,4 @@
 // (C) 2007-2020 GoodData Corporation
 
+export { clearInsightViewCaches } from "./dataLoaders";
 export * from "./insightView";
