+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/hooks/useDashboard.ts
@@ -10,7 +10,7 @@ import {
 } from "@gooddata/sdk-ui";
 import { ObjRef, objRefToString } from "@gooddata/sdk-model";
 import invariant from "ts-invariant";
-import { getDashboardViewDataLoader } from "./dataLoaders";
+import { dashboardDataLoaderFactory } from "./dataLoaders";
 
 /**
  * @beta
@@ -66,7 +66,7 @@ export function useDashboard({
         "The workspace in useDashboard must be defined. Either pass it as a config prop or make sure there is a WorkspaceProvider up the component tree.",
     );
 
-    const loader = getDashboardViewDataLoader(effectiveWorkspace);
+    const loader = dashboardDataLoaderFactory.forWorkspace(effectiveWorkspace);
     const promise = () => loader.getDashboard(effectiveBackend, dashboard);
 
     return useCancelablePromise({ promise, onCancel, onError, onLoading, onPending, onSuccess }, [
