+++ libs/sdk-ui-ext/src/internal/dashboardEmbedding/hooks/useUserWorkspacePermissions.ts
@@ -9,7 +9,7 @@ import {
     useWorkspace,
 } from "@gooddata/sdk-ui";
 import invariant from "ts-invariant";
-import { getDashboardViewDataLoader } from "./dataLoaders";
+import { userWorkspacePermissionsDataLoaderFactory } from "./dataLoaders";
 
 /**
  * @beta
@@ -60,7 +60,7 @@ export function useUserWorkspacePermissions({
         "The workspace in useUserWorkspacePermissions must be defined. Either pass it as a config prop or make sure there is a WorkspaceProvider up the component tree.",
     );
 
-    const loader = getDashboardViewDataLoader(effectiveWorkspace);
+    const loader = userWorkspacePermissionsDataLoaderFactory.forWorkspace(effectiveWorkspace);
     const promise = () => loader.getUserWorkspacePermissions(effectiveBackend);
 
     return useCancelablePromise({ promise, onCancel, onError, onLoading, onPending, onSuccess }, [
